outsource storage function to tardis_minimal_model.py
