+++ tardis_kromer_plot.py
@@ -28,126 +28,6 @@ elements = {'neut': 0, 'h': 1, 'he': 2, 'li': 3, 'be': 4, 'b': 5, 'c': 6, 'n':
 inv_elements = dict([(v, k) for k, v in elements.items()])
 
 
-def store_data_for_kromer_plot(mdl, buffer_or_fname="minimal_model.hdf5",
-                               path="", mode="virtual"):
-    """Simple helper routine to dump all information required to generate a
-    kromer plot to hdf5 files.
-
-    Parameters
-    ----------
-    mdl : tardis.model.Radial1DModel
-        source tardis model object
-    lines_fname : str
-        name of the hdf5 file for the line list (default 'lines.hdf5')
-    model_fname : str
-        name of the hdf5 file for the model data (default 'model.hdf5')
-    """
-
-    def _save_spectrum_real(key, path, hdf_store):
-
-        wave = mdl.spectrum.wavelength.value
-        flux = mdl.spectrum.luminosity_density_lambda.value
-
-        luminosity_density = \
-            pd.DataFrame.from_dict(dict(wave=wave, flux=flux))
-        luminosity_density.to_hdf(hdf_store, os.path.join(path, key))
-
-    def _save_spectrum_virtual(key, path, hdf_store):
-
-        wave = mdl.spectrum_virtual.wavelength.value
-        flux = mdl.spectrum_virtual.luminosity_density_lambda.value
-
-        luminosity_density_virtual = pd.DataFrame.from_dict(dict(wave=wave,
-                                                                 flux=flux))
-        luminosity_density_virtual.to_hdf(hdf_store, os.path.join(path, key))
-
-    def _save_configuration_dict(key, path, hdf_store):
-        configuration_dict = dict(time_of_simulation=mdl.time_of_simulation)
-        configuration_dict_path = os.path.join(path, 'configuration')
-        pd.Series(configuration_dict).to_hdf(hdf_store,
-                                             configuration_dict_path)
-
-    if mode == "virtual":
-        include_from_runner_ = \
-            {'virt_packet_last_interaction_type': None,
-             'virt_packet_last_line_interaction_in_id': None,
-             'virt_packet_last_line_interaction_out_id': None,
-             'virt_packet_last_interaction_in_nu': None,
-             'virt_packet_nus': None,
-             'virt_packet_energies': None}
-        include_from_spectrum_ = \
-            {'luminosity_density_virtual': _save_spectrum_virtual}
-    elif mode == "real":
-        include_from_runner_ = \
-            {'last_interaction_type': None,
-             'last_line_interaction_in_id': None,
-             'last_line_interaction_out_id': None,
-             'last_interaction_in_nu': None,
-             'output_nu': None,
-             'output_energy': None}
-        include_from_spectrum_ = \
-            {'luminosity_density': _save_spectrum_real}
-
-    else:
-        raise ValueError
-
-    include_from_atom_data_ = {'lines': None}
-    include_from_model_in_hdf5 = {'runner': include_from_runner_,
-                                  'atom_data': include_from_atom_data_,
-                                  'spectrum': include_from_spectrum_,
-                                  'configuration_dict':
-                                  _save_configuration_dict,
-                                  }
-
-    if isinstance(buffer_or_fname, basestring):
-        hdf_store = pd.HDFStore(buffer_or_fname)
-    elif isinstance(buffer_or_fname, pd.HDFStore):
-        hdf_store = buffer_or_fname
-    else:
-        raise IOError('Please specify either a filename or an HDFStore')
-    logger.info('Writing to path %s', path)
-
-    def _get_hdf5_path(path, property_name):
-        return os.path.join(path, property_name)
-
-    def _to_smallest_pandas(object):
-        try:
-            return pd.Series(object)
-        except Exception:
-            return pd.DataFrame(object)
-
-    def _save_model_property(object, property_name, path, hdf_store):
-        property_path = _get_hdf5_path(path, property_name)
-
-        try:
-            object.to_hdf(hdf_store, property_path)
-        except AttributeError:
-            _to_smallest_pandas(object).to_hdf(hdf_store, property_path)
-
-    for key in include_from_model_in_hdf5:
-        if include_from_model_in_hdf5[key] is None:
-            _save_model_property(getattr(mdl, key), key, path, hdf_store)
-        elif callable(include_from_model_in_hdf5[key]):
-            include_from_model_in_hdf5[key](key, path, hdf_store)
-        else:
-            try:
-                for subkey in include_from_model_in_hdf5[key]:
-                    if include_from_model_in_hdf5[key][subkey] is None:
-                        _save_model_property(getattr(getattr(mdl, key),
-                                                     subkey), subkey,
-                                             os.path.join(path, key),
-                                             hdf_store)
-                    elif callable(include_from_model_in_hdf5[key][subkey]):
-                        include_from_model_in_hdf5[key][subkey](subkey, os.path.join(path, key), hdf_store)
-                    else:
-                        logger.critical('Can not save %s',
-                                        str(os.path.join(path, key, subkey)))
-            except:
-                logger.critical('An error occurred while dumping %s to HDF.',
-                                str(os.path.join(path, key)))
-
-    hdf_store.flush()
-    hdf_store.close()
 
 
 class tardis_kromer_plotter(object):
